done enough to pass part7

+++ part7/bloglist/src/App.js
@@ -1,24 +1,28 @@
-import React, { useState, useEffect } from 'react'
+import React, { useEffect } from 'react'
+import { BrowserRouter as Router, Route } from 'react-router-dom'
 import { connect } from 'react-redux'
+import { useField } from './hooks'
+import { setNotification, removeNotification } from './reducers/notificationReducer'
+import { initBlogs, updateBlog, addBlog, removeBlog } from './reducers/blogReducer'
+import { tryLogin, setLogin, loginOut } from './reducers/loginReducer'
+
 import Blog from './components/Blog'
 import blogService from './services/blogs'
-import loginService from './services/login'
 import NewBlog from './components/NewBlog'
 import Notification from './components/Notification'
 import Togglable from './components/Togglable'
-import { useField } from './hooks'
-import { setNotification } from './reducers/notificationReducer'
-import { initBlogs, updateBlog, addBlog, removeBlog } from './reducers/blogReducer'
-import { tryLogin, setLogin } from './reducers/loginReducer'
+import Users from './components/Users'
+import User from './components/User'
 
 
 const App = (props) => {
   const [username] = useField('text')
   const [password] = useField('password')
-  // const [user, setUser] = useState(null)
 
   useEffect(() => {
-    props.initBlogs()
+    if(props.login) {
+      props.initBlogs()
+    }
     // eslint-disable-next-line
   }, [])
 
@@ -29,49 +33,59 @@ const App = (props) => {
       props.setLogin(user)
       blogService.setToken(user.token)
     }
+    // eslint-disable-next-line
   }, [])
 
+  useEffect(() => {
+    if(props.login) {
+      window.localStorage.setItem('loggedBlogAppUser', JSON.stringify(props.login))
+      blogService.setToken(props.login.token)
+      props.initBlogs()
+    } else {
+      blogService.destroyToken()
+      window.localStorage.removeItem('loggedBlogAppUser')
+      notify(null)
+    }
+    // eslint-disable-next-line
+  }, [props.login])
+
   const notify = (message, type = 'success') => {
-    props.setNotification({ message, type },  5)
+    if(message === null) {
+      props.removeNotification()
+    } else {
+      props.setNotification({ message, type },  5)
+    }
   }
 
   const handleLogin = async (event) => {
     event.preventDefault()
     try {
-
-      const user = props
-      const user = await loginService.login({
+      await props.tryLogin({
         username: username.value,
         password: password.value
       })
-
-      window.localStorage.setItem('loggedBlogAppUser', JSON.stringify(user))
-      blogService.setToken(user.token)
-      setUser(user)
     } catch (exception) {
       notify('wrong username of password', 'error')
     }
   }
 
   const handleLogout = () => {
-    setUser(null)
-    blogService.destroyToken()
-    window.localStorage.removeItem('loggedBlogAppUser')
+    props.loginOut()
   }
 
-  const createBlog = async (blog) => {
+  const createBlog = (blog) => {
     props.addBlog(blog)
     newBlogRef.current.toggleVisibility()
     notify(`a new blog ${blog.title} by ${blog.author} added`)
   }
 
-  const likeBlog = async (blog) => {
+  const likeBlog = (blog) => {
     const likedBlog = { ...blog, likes: blog.likes + 1 }
     props.updateBlog(likedBlog)
     notify(`blog ${likedBlog.title} by ${likedBlog.author} liked!`)
   }
 
-  const removeBlog = async (blog) => {
+  const removeBlog = (blog) => {
     const ok = window.confirm(`remove blog ${blog.title} by ${blog.author}`)
     if (ok) {
       props.removeBlog(blog)
@@ -79,7 +93,7 @@ const App = (props) => {
     }
   }
 
-  if (user === null) {
+  if (props.login === null) {
     return (
       <div>
         <h2>log in to application</h2>
@@ -95,7 +109,7 @@ const App = (props) => {
             password
             <input {...password} />
           </div>
-          <button type="submit">kirjaudu</button>
+          <button type="submit">login</button>
         </form>
       </div>
     )
@@ -103,31 +117,51 @@ const App = (props) => {
 
   const newBlogRef = React.createRef()
 
+  const blogListRender = () => {
+    return (
+      <>
+        <Togglable buttonLabel='create new' ref={newBlogRef}>
+          <NewBlog createBlog={createBlog} />
+        </Togglable>
+
+        {props.visibleBlogs.map(blog => {
+          return <Blog
+            key={blog.id}
+            blog={blog}
+            like={likeBlog}
+            remove={removeBlog}
+            user={props.login}
+            creator={blog.user.username === props.login.username}
+          />
+        }
+        )}
+      </>
+    )
+  }
+
+  const blogsByUser = (index) => {
+    return props.visibleBlogs.filter(blog => {
+      return blog.user.id === index
+    })
+  }
+
   return (
-    <div>
-      <h2>blogs</h2>
+    <Router>
+      <div>
+        <h2>blogs</h2>
 
-      <Notification/>
+        <Notification/>
 
-      <p>{user.name} logged in</p>
-      <button onClick={handleLogout}>logout</button>
+        <p>{props.login.name} logged in</p>
+        <button onClick={handleLogout}>logout</button>
 
-      <Togglable buttonLabel='create new' ref={newBlogRef}>
-        <NewBlog createBlog={createBlog} />
-      </Togglable>
-
-      {props.visibleBlogs.map(blog => {
-        return <Blog
-          key={blog.id}
-          blog={blog}
-          like={likeBlog}
-          remove={removeBlog}
-          user={user}
-          creator={blog.user.username === user.username}
+        <Route exact path="/" render={() => blogListRender()} />
+        <Route exact path="/users" render={() => <Users/>} />
+        <Route exact path="/users/:id"
+          render={({ match }) => <User blogs={blogsByUser(match.params.id)} />}
         />
-      }
-      )}
-    </div>
+      </div>
+    </Router>
   )
 }
 
@@ -135,26 +169,27 @@ const blogsToShow = ({ blogs }) => {
   if(!blogs || blogs.length === 0) {
     return []
   }
-
   const byLikes = (b1, b2) => b2.likes - b1.likes
-
   return blogs.sort(byLikes)
 }
 
 const mapStateToProps = (state) => {
   return {
-    visibleBlogs: blogsToShow(state)
+    visibleBlogs: blogsToShow(state),
+    login: state.login
   }
 }
 
 const mapDispatchToProps = {
   initBlogs,
   setNotification,
+  removeNotification,
   updateBlog,
   addBlog,
   removeBlog,
   tryLogin,
   setLogin,
+  loginOut,
 }
 
 const ConnectedApp = connect(
